safeint for region bytes in bfc arena and code clean up  (#3447)
* PR comments

* remove build issue workaround

* SafeInt for region bytes

* fix build

* fix build

+++ onnxruntime/core/framework/bfc_arena.cc
@@ -99,9 +99,10 @@ bool BFCArena::Extend(size_t rounded_bytes) {
         increased_allocation = true;
       }
 
-      extend_bytes = std::min(curr_region_allocation_bytes_, available_bytes);
+      extend_bytes = std::min(static_cast<size_t>(curr_region_allocation_bytes_), available_bytes);
 
-      // we allocated the same number of bytes as the current region, so we have 2x that now
+      // we allocated the same number of bytes as the current region
+      // the 2x is to double the minimum size of the next amount we'll allocate
       if (!increased_allocation) {
         curr_region_allocation_bytes_ *= 2;
       }
